[xslocobot_landmark_nav] Added new example pkg
The xslocobot_landmark_nav example package builds the navigation
package and the landmark modules to allow the user to go to points of
interest on the map(marked with an AR tag). The map must first be
explored and filled out using rtabmap and landmark_finder. Then, the
user may navigation to the AR tag location using the nav_to_landmark
script.